vm-apiZoho/requirements.txt
requests==2.31.0
fastapi==0.110.2
loguru==0.7.2
aiohttp==3.8.5
uvicorn==0.29.0
mysql-connector-python==8.0.33
